Use a group retry algorithm to fix a lock request retry race which would cause a pending lock request to stay pending and eventually time out after the conflicting lock has been released.
The race in the old algorithm occurs when thread A releases its locks and then retries the
pending lock requests.  If thread B releases its locks and then tries to retry the pending lock
requests, it gets pushed away since thread A is running lock retry.  Thread A may have retried
the locks that thread B released BEFORE thread B released them.  If this is the case, when thread A
finishes lock retry, there still will be a pending lock request with no lock conflicts.

The new algorithm ensures that all pending lock requests are retried for all released locks. In
addition, the new algorithm groups retries together so that a minimal amount of work is done.